refactor: move construct table func to output.go
