+++ app/app.go
@@ -2,7 +2,6 @@ package app
 
 import (
 	"fmt"
-	"github.com/clinaresl/table"
 	"io/fs"
 	"os"
 	"os/exec"
@@ -48,53 +47,6 @@ func GetRepos(root string) ([]Repo, error) {
 		return !repo.valid
 	})
 	return repos, nil
-}
-
-func ConstructSummary(repos []Repo, root string) string {
-	countRepos := len(repos)
-	var countUnsynced int
-	for _, repo := range repos {
-		if !repo.SyncedWithRemote {
-			countUnsynced++
-		}
-	}
-	return fmt.Sprintf(
-		"%v repos found in %v: %v repo(s) are not synced",
-		countRepos,
-		root,
-		countUnsynced,
-	)
-}
-
-func ConstructTable(repos []Repo) (*table.Table, error) {
-	t, err := table.NewTable("| C{15} | L{20} | c | c | L{27} |")
-	if err != nil {
-		return nil, err
-	}
-	t.AddThickRule()
-	t.AddRow("Repo", "Path", "Last Modified", "Synced?", "Sync Details")
-	t.AddThickRule()
-	for i := range repos {
-		year, month, day := repos[i].LastModified.Date()
-		LastModifiedDate := fmt.Sprintf("%04d-%02d-%02d", year, int(month), day)
-		var isSynced string
-		if repos[i].SyncedWithRemote {
-			isSynced = "yes"
-		} else {
-			isSynced = "no"
-		}
-
-		t.AddRow(
-			repos[i].Name,
-			repos[i].AbsPath,
-			LastModifiedDate,
-			isSynced,
-			repos[i].SyncDescription,
-		)
-		t.AddSingleRule()
-	}
-	return t, nil
-
 }
 
 func evaluateCommitSyncStatus(gitOut string) (bool, string) {
